Add serde target
