+++ fuzz/Cargo.toml
@@ -10,9 +10,11 @@ cargo-fuzz = true
 [dependencies]
 arbitrary = { version = "1", features = ["derive"] }
 libfuzzer-sys = "0.4"
+serde_json = "1.0.96"
 
 [dependencies.petgraph]
 path = ".."
+features = ["serde-1"]
 
 # Prevent this from interfering with workspaces
 [workspace]
@@ -21,8 +23,18 @@ members = ["."]
 [profile.release]
 debug = 1
 
+[lib]
+name = "util"
+path = "fuzz_targets/util.rs"
+
 [[bin]]
 name = "mst"
 path = "fuzz_targets/mst.rs"
+test = false
+doc = false
+
+[[bin]]
+name = "serde"
+path = "fuzz_targets/serde.rs"
 test = false
 doc = false
